moving some classes to different files
