+++ src/SayedHa.Blackjack.Shared/Blackjack/Strategy/Tree/CardNumberOrScore.cs
@@ -1,0 +1,38 @@
+﻿namespace SayedHa.Blackjack.Shared.Blackjack.Strategy.Tree {
+    public enum CardNumberOrScore {
+        Ace,
+        Two,
+        Three,
+        Four,
+        Five,
+        Six,
+        Seven,
+        Eight,
+        Nine,
+        Ten,
+        Jack,
+        Queen,
+        King,
+        Score21,
+        Score20,
+        Score19,
+        Score18,
+        Score17,
+        Score16,
+        Score15,
+        Score14,
+        Score13,
+        Score12,
+        Score11,
+        Score10,
+        Score9,
+        Score8,
+        Score7,
+        Score6,
+        Score5,
+        Score4,
+        Score3,
+        Score2,
+        Busted
+    }
+}
