+++ src/SayedHa.Blackjack.Shared/Blackjack/Strategy/Tree/sample.csv
@@ -1,0 +1,1 @@
+﻿2,17,Stand
